🐛 [Fix] fix errors caused by incorrect data binding in category filter view
- 카테고리 메인 > 1차 필터뷰로 이동 시 d1CateogryCode 바인딩 에러로 발생하던 api 에러 해결
- 주석 및 폴더 정리
- 추가 사항 : 1Depth 이동 시 Category 필터 수정 필요
