+++ Projects/App/Sources/Data/DTO/Filter/D2CategoryFilterResponseDTO.swift
