+++ Projects/App/Sources/Data/DTO/Filter/ZeroCategoryFilterResponseDTO.swift
