+++ Projects/App/Sources/Presentation/Categoery/Filter/CategoryListView.swift
@@ -91,7 +91,3 @@ struct CategoryListView: View {
         }
     }
 }
-
-//#Preview {
-//    CategoryListView(viewModel: CategoryFilteredViewModel(category: ""))
-//}
